+++ src/components/Popover/Popover.js
@@ -3,16 +3,15 @@ import PropTypes from 'prop-types';
 import cx from 'classnames';
 import omit from 'lodash.omit';
 
-import { themeable } from '../../decorators';
-
 import { Button, ButtonIcon } from '../../';
 
 export class Popover extends React.Component {
-  static getThemeName(themeStr) {
+  static shouldInvertIcon(themeStr) {
     if (typeof themeStr === 'string') {
       return themeStr.includes('error') ||
         themeStr.includes('success') ||
-        themeStr.includes('info');
+        themeStr.includes('info') ||
+        themeStr.includes('offline');
     }
     return false;
   }
@@ -22,18 +21,18 @@ export class Popover extends React.Component {
   }
 
   renderHeader() {
-    const { header, panels, customLayout } = this.props;
+    const { header, panels, customHeaderTheme } = this.props;
     const headerClasses = [
       'slds-popover__header',
-      { 'slds-theme_warning': customLayout === 'warning' },
-      { 'slds-theme_error': customLayout === 'error' },
-      { 'slds-theme_success': customLayout === 'success' },
-      { 'slds-theme_info': customLayout === 'info' },
+      { 'slds-theme_warning': customHeaderTheme === 'warning' },
+      { 'slds-theme_error': customHeaderTheme === 'error' },
+      { 'slds-theme_success': customHeaderTheme === 'success' },
+      { 'slds-theme_info': customHeaderTheme === 'info' },
     ];
 
     let headerContent;
     let borderRadius;
-    if (!!panels && (typeof customLayout !== 'undefined') && customLayout !== '') {
+    if (!!panels && (typeof customHeaderTheme !== 'undefined') && customHeaderTheme !== '') {
       headerContent = (<h2 className="slds-text-heading_small slds-p-around_xxx-small">{header}</h2>);
       borderRadius = {
         borderTopLeftRadius: 'calc(0.25rem - 1px)',
@@ -51,13 +50,10 @@ export class Popover extends React.Component {
   }
 
   renderBody() {
-    const { body, customLayout } = this.props;
-    const bodyClasses = [
-      { 'slds-popover__body': (typeof customLayout === 'undefined' || customLayout === '') },
-    ];
+    const { body } = this.props;
 
     return (
-      <div className={cx(bodyClasses)}>
+      <div className="slds-popover__body">
         {body}
       </div>
     );
@@ -74,8 +70,9 @@ export class Popover extends React.Component {
   }
 
   renderCloseButton() {
-    const { onClose, className, customLayout } = this.props;
-    const invertIcon = customLayout ? Popover.getThemeName(customLayout) : Popover.getThemeName(className);
+    const { onClose, customHeaderTheme, theme } = this.props;
+    const invertIcon =
+      customHeaderTheme ? Popover.shouldInvertIcon(customHeaderTheme) : Popover.shouldInvertIcon(theme);
     const closeButtonClasses = [
       'slds-button_icon-small',
       'slds-float_right',
@@ -96,14 +93,15 @@ export class Popover extends React.Component {
   }
 
   render() {
-    const { className, closeable, open, customLayout, nubbin, panels, header, body, footer } = this.props;
+    const { className, closeable, open, customHeaderTheme, nubbin, panels, header, body, footer, theme } = this.props;
     const rest = omit(this.props, Object.keys(Popover.propTypes));
 
     const sldsClasses = [
       'slds-popover',
       { [`slds-nubbin_${nubbin}`]: !!nubbin },
-      { 'slds-popover_panel': (!!panels && (typeof customLayout === 'undefined' || customLayout === '')) },
+      { 'slds-popover_panel': (!!panels && (typeof customHeaderTheme === 'undefined' || customHeaderTheme === '')) },
       { 'slds-hide': !open },
+      { [`slds-theme_${theme}`]: !!theme },
       className,
     ];
 
@@ -127,11 +125,12 @@ Popover.defaultProps = {
   header: null,
   footer: null,
   className: null,
-  customLayout: undefined,
+  customHeaderTheme: undefined,
   open: false,
   closeable: true,
   panels: false,
   nubbin: 'bottom-left',
+  theme: null,
 };
 
 Popover.propTypes = {
@@ -150,24 +149,15 @@ Popover.propTypes = {
   /**
    * Popover header content
    */
-  header: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.node,
-  ]),
+  header: PropTypes.node,
   /**
    * Popover body content
    */
-  body: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.node,
-  ]),
+  body: PropTypes.node,
   /**
    * Popover footer content
    */
-  footer: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.node,
-  ]),
+  footer: PropTypes.node,
   /**
    * Additional css classes
    */
@@ -179,11 +169,39 @@ Popover.propTypes = {
   /**
    * Optional position of nubbin
    */
-  nubbin: PropTypes.string,
+  nubbin: PropTypes.oneOf([
+    'left',
+    'left-top',
+    'left-bottom',
+    'top-left',
+    'top-right',
+    'right-top',
+    'right-bottom',
+    'bottom-left',
+    'bottom-right',
+  ]),
   /**
-   * Optional custom layout (warning, error, success, info)
+   * Optional custom layout
    */
-  customLayout: PropTypes.string,
+  customHeaderTheme: PropTypes.oneOf([
+    'warning',
+    'error',
+    'success',
+    'info']),
+  /**
+   * theme
+   */
+  theme: PropTypes.oneOf([
+    'alt-inverse',
+    'default',
+    'error',
+    'info',
+    'inverse',
+    'offline',
+    'shade',
+    'success',
+    'warning',
+  ]),
 };
 
-export default themeable(Popover);
+export default Popover;